Update Navigation Button Styling
