+++ src/Components/Navigation/Navigation.js
@@ -7,7 +7,6 @@ import MuiAppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import List from '@mui/material/List';
 import Typography from '@mui/material/Typography';
-import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
 import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
@@ -17,7 +16,7 @@ import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import SearchIcon from '@mui/icons-material/Search';
 import { DisplayItems, ProfileTab } from './ListItems';
-import { NavigationContext, useNavigationContext } from '../../Context/NavigationProvider'
+import { useNavigationContext } from '../../Context/NavigationProvider'
 
 const User = {
     name: 'Jay Nayon',
