+++ src/Components/Navigation/listItems.js
@@ -3,73 +3,14 @@ import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
 import DashboardIcon from '@mui/icons-material/Dashboard';
-import SchoolIcon from '@mui/icons-material/School';
 import PeopleIcon from '@mui/icons-material/People';
 import BarChartIcon from '@mui/icons-material/BarChart';
 import LogoutIcon from '@mui/icons-material/Logout';
 import SettingsIcon from '@mui/icons-material/Settings'
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import Divider from '@mui/material/Divider';
-import Collapse from "@mui/material/Collapse";
-import ExpandLess from "@mui/icons-material/ExpandLess";
-import ExpandMore from "@mui/icons-material/ExpandMore";
-import List from '@mui/material/List';
 import { Link } from 'react-router-dom';
-
-function DisplaySchools({ selected, setSelected }) {
-    const User = {
-        name: 'Jay Nayon',
-        email: 'jay.nayonjr@cit.edu',
-        schools: [
-            'Jaclupan ES',
-            'Talisay ES'
-        ]
-    }
-
-    const [open, setOpen] = useState(false);
-
-    const handleClick = () => {
-        setOpen(!open);
-    };
-    return (
-        <React.Fragment>
-            <ListItemButton sx={styles.button} onClick={handleClick} >
-                <ListItemIcon sx={{ width: 'auto', minWidth: '40px' }}>
-                    <SchoolIcon sx={styles.icon} />
-                </ListItemIcon>
-                <ListItemText
-                    primary={"School"}
-                    primaryTypographyProps={open ? styles.typography.school : styles.typography}
-                />
-                {open ? <ExpandLess sx={{ color: 'white' }} /> : <ExpandMore sx={{ color: 'white' }} />}
-            </ListItemButton>
-            <Collapse in={open} timeout="auto" unmountOnExit >
-                <div style={{ display: 'flex' }}>
-                    <VerticalLine width='60px' color='white' />
-                    <List component="div" disablePadding>
-                        {User.schools.map((item, index) => {
-                            return (
-                                <ListItemButton
-                                    key={index}
-                                    //component={Link}
-                                    //to={index < 5 ? `/${item}` : '/'}
-                                    sx={styles.button}
-                                    value={item}
-                                    selected={selected === item}
-                                    onClick={() => { setSelected(item) }}
-                                >
-                                    <ListItemText
-                                        primary={item}
-                                        primaryTypographyProps={styles.typography} />
-                                </ListItemButton>
-                            )
-                        })}
-                    </List>
-                </div>
-            </Collapse>
-        </React.Fragment>
-    )
-}
+import DisplaySchools from './DisplaySchools';
 
 export function DisplayItems() {
     const list = ['dashboard', 'schools', 'people', 'settings', 'testing', 'logout'];
@@ -158,7 +99,7 @@ export const ProfileTab = ({ user }) => {
     );
 };
 
-function VerticalLine({ width, color = 'black' }) {
+export function VerticalLine({ width, color = 'black' }) {
     return (
         <div style={{
             display: 'flex',
@@ -175,7 +116,7 @@ function VerticalLine({ width, color = 'black' }) {
     );
 }
 
-const styles = {
+export const styles = {
     icon: {
         color: 'white',
         fontSize: '19px',
@@ -202,15 +143,16 @@ const styles = {
     },
     button: {
         "&.Mui-selected": {
-            backgroundColor: 'rgba(255, 255, 255, 0.2)', // Change to desired highlight color
+            backgroundColor: 'rgba(255, 255, 255, 0.10)', // Change to desired highlight color
             borderRadius: '10px',
         },
         "& .MuiTouchRipple-root": {
             color: 'white'
         },
-        /*'&:hover, &.Mui-focusVisible': {
-            backgroundColor: 'white'
-        },*/
+        '&:hover, &.Mui-focusVisible': {
+            backgroundColor: 'rgba(255, 255, 255, 0.05)',
+            borderRadius: '10px',
+        },
         marginRight: '5px',
         marginLeft: '5px',
         paddingTop: '5px',
